- fixed the broken integration test by changing the pojo to use double instead of big decimal, as the latter is not supported by BSON - upgraded jongo to 0.4

+++ src/main/org/monkey/sample/model/SampleDailyPrice.java
@@ -3,22 +3,20 @@ package org.monkey.sample.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.joda.time.LocalDate;
 
-import java.math.BigDecimal;
-
 public class SampleDailyPrice {
 
     @JsonProperty private String ricCode;
     @JsonProperty private LocalDate tradeDate;
-    @JsonProperty private BigDecimal open;
-    @JsonProperty private BigDecimal high;
-    @JsonProperty private BigDecimal low;
-    @JsonProperty private BigDecimal close;
-    @JsonProperty private BigDecimal adjClose;
+    @JsonProperty private double open;
+    @JsonProperty private double high;
+    @JsonProperty private double low;
+    @JsonProperty private double close;
+    @JsonProperty private double adjClose;
 
     private SampleDailyPrice() {
     }
 
-    public SampleDailyPrice(String ricCode, LocalDate tradeDate, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close, BigDecimal adjClose) {
+    public SampleDailyPrice(String ricCode, LocalDate tradeDate, double open, double high, double low, double close, double adjClose) {
         this.ricCode = ricCode;
         this.tradeDate = tradeDate;
         this.open = open;
@@ -36,23 +34,23 @@ public class SampleDailyPrice {
         return tradeDate;
     }
 
-    public BigDecimal getOpen() {
+    public double getOpen() {
         return open;
     }
 
-    public BigDecimal getHigh() {
+    public double getHigh() {
         return high;
     }
 
-    public BigDecimal getLow() {
+    public double getLow() {
         return low;
     }
 
-    public BigDecimal getClose() {
+    public double getClose() {
         return close;
     }
 
-    public BigDecimal getAdjClose() {
+    public double getAdjClose() {
         return adjClose;
     }
 }
